Add requirements.txt and configuration to deploy in vercel
