+++ flaskApi/requirements.txt
@@ -1,0 +1,5 @@
+flask == 2.3.2
+tensorflow == 2.12.0
+numpy == 1.23.2
+psycopg2 == 2.9.1
+cv2 == 4.5.3
